separate metrics.rs into multiple files
