+++ chain-signatures/node/src/metrics/storage.rs
@@ -1,0 +1,51 @@
+use std::sync::LazyLock;
+
+use prometheus::{exponential_buckets, HistogramVec, IntGaugeVec};
+
+use super::{try_create_histogram_vec, try_create_int_gauge_vec};
+
+pub(crate) static REDIS_LATENCY: LazyLock<HistogramVec> = LazyLock::new(|| {
+    try_create_histogram_vec(
+        "multichain_redis_operation_latency_ms",
+        "Latency of Redis operations in storage layers",
+        &["protocol", "operation", "node_account_id"],
+        Some(exponential_buckets(1.0, 2.0, 15).unwrap()),
+    )
+    .unwrap()
+});
+
+pub(crate) static NUM_TRIPLES_MINE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
+    try_create_int_gauge_vec(
+        "multichain_num_triples_mine",
+        "number of triples of the node's own",
+        &["node_account_id"],
+    )
+    .unwrap()
+});
+
+pub(crate) static NUM_TRIPLES_TOTAL: LazyLock<IntGaugeVec> = LazyLock::new(|| {
+    try_create_int_gauge_vec(
+        "multichain_num_triples_total",
+        "number of total triples",
+        &["node_account_id"],
+    )
+    .unwrap()
+});
+
+pub(crate) static NUM_PRESIGNATURES_MINE: LazyLock<IntGaugeVec> = LazyLock::new(|| {
+    try_create_int_gauge_vec(
+        "multichain_num_presignatures_mine",
+        "number of presignatures of the node's own",
+        &["node_account_id"],
+    )
+    .unwrap()
+});
+
+pub(crate) static NUM_PRESIGNATURES_TOTAL: LazyLock<IntGaugeVec> = LazyLock::new(|| {
+    try_create_int_gauge_vec(
+        "multichain_num_presignatures_total",
+        "number of total presignatures",
+        &["node_account_id"],
+    )
+    .unwrap()
+});
