+++ chain-signatures/node/src/protocol/triple.rs
@@ -112,25 +112,26 @@ impl TripleGenerator {
     }
 
     async fn run(mut self, my_account_id: AccountId, epoch: u64) {
-        let before_first_poke_delay =
-            crate::metrics::TRIPLE_BEFORE_POKE_DELAY.with_label_values(&[my_account_id.as_str()]);
-        let accrued_wait_delay =
-            crate::metrics::TRIPLE_ACCRUED_WAIT_DELAY.with_label_values(&[my_account_id.as_str()]);
-        let runtime_latency =
-            crate::metrics::TRIPLE_LATENCY.with_label_values(&[my_account_id.as_str()]);
-        let total_latency =
-            crate::metrics::TRIPLE_LATENCY_TOTAL.with_label_values(&[my_account_id.as_str()]);
-        let poke_latency =
-            crate::metrics::TRIPLE_POKE_CPU_TIME.with_label_values(&[my_account_id.as_str()]);
-        let poke_counts =
-            crate::metrics::TRIPLE_POKES_CNT.with_label_values(&[my_account_id.as_str()]);
-        let success_owned_counts =
-            crate::metrics::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATIONS_MINE_SUCCESS
-                .with_label_values(&[my_account_id.as_str()]);
-        let success_total_counts = crate::metrics::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATORS_SUCCESS
+        let before_first_poke_delay = crate::metrics::protocols::TRIPLE_BEFORE_POKE_DELAY
             .with_label_values(&[my_account_id.as_str()]);
-        let failure_counts =
-            crate::metrics::TRIPLE_GENERATOR_FAILURES.with_label_values(&[my_account_id.as_str()]);
+        let accrued_wait_delay = crate::metrics::protocols::TRIPLE_ACCRUED_WAIT_DELAY
+            .with_label_values(&[my_account_id.as_str()]);
+        let runtime_latency =
+            crate::metrics::protocols::TRIPLE_LATENCY.with_label_values(&[my_account_id.as_str()]);
+        let total_latency = crate::metrics::protocols::TRIPLE_LATENCY_TOTAL
+            .with_label_values(&[my_account_id.as_str()]);
+        let poke_latency = crate::metrics::protocols::TRIPLE_POKE_CPU_TIME
+            .with_label_values(&[my_account_id.as_str()]);
+        let poke_counts = crate::metrics::protocols::TRIPLE_POKES_CNT
+            .with_label_values(&[my_account_id.as_str()]);
+        let success_owned_counts =
+            crate::metrics::protocols::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATIONS_MINE_SUCCESS
+                .with_label_values(&[my_account_id.as_str()]);
+        let success_total_counts =
+            crate::metrics::protocols::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATORS_SUCCESS
+                .with_label_values(&[my_account_id.as_str()]);
+        let failure_counts = crate::metrics::protocols::TRIPLE_GENERATOR_FAILURES
+            .with_label_values(&[my_account_id.as_str()]);
 
         let start_time = Instant::now();
         let mut total_wait = Duration::from_millis(0);
@@ -503,7 +504,7 @@ impl TripleSpawner {
 
         self.ongoing
             .spawn(id, generator.run(self.my_account_id.clone(), self.epoch));
-        crate::metrics::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATORS
+        crate::metrics::protocols::NUM_TOTAL_HISTORICAL_TRIPLE_GENERATORS
             .with_label_values(&[self.my_account_id.as_str()])
             .inc();
 
@@ -580,16 +581,16 @@ impl TripleSpawner {
                     self.stockpile(&active, &protocol).await;
                     let _ = ongoing_gen_tx.send(self.ongoing.len());
 
-                    crate::metrics::NUM_TRIPLES_MINE
+                    crate::metrics::storage::NUM_TRIPLES_MINE
                         .with_label_values(&[self.my_account_id.as_str()])
                         .set(self.len_mine().await as i64);
-                    crate::metrics::NUM_TRIPLES_TOTAL
+                    crate::metrics::storage::NUM_TRIPLES_TOTAL
                         .with_label_values(&[self.my_account_id.as_str()])
                         .set(self.triple_storage.len_generated().await as i64);
-                    crate::metrics::NUM_TRIPLE_GENERATORS_INTRODUCED
+                    crate::metrics::protocols::NUM_TRIPLE_GENERATORS_INTRODUCED
                         .with_label_values(&[self.my_account_id.as_str()])
                         .set(self.len_introduced() as i64);
-                    crate::metrics::NUM_TRIPLE_GENERATORS_TOTAL
+                    crate::metrics::protocols::NUM_TRIPLE_GENERATORS_TOTAL
                         .with_label_values(&[self.my_account_id.as_str()])
                         .set(self.len_ongoing() as i64);
                 }
